Fix histogram bucket overflow: u32 -> u64
Histogram buckets were uint32, which overflows in ~12 hours at
100K events/sec. Change to uint64 across BPF map struct, userspace
accumulator, per-CPU summation, and output formatting.

+++ src/map_reader.c
@@ -137,7 +137,7 @@ int pgwt_read_maps(struct pgwt_daemon *d)
 
         /* Sum across all CPUs */
         uint64_t count = 0, total = 0, min_v = UINT64_MAX, max_v = 0;
-        uint32_t hist[HISTOGRAM_BUCKETS] = {};
+        uint64_t hist[HISTOGRAM_BUCKETS] = {};
 
         for (int cpu = 0; cpu < nr_cpus && cpu < MAX_CPUS; cpu++) {
             count += values[cpu].count;
